% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carlotti_and_Parast_2026_simulations.R
\name{Carlotti_and_Parast_2026_simulations}
\alias{Carlotti_and_Parast_2026_simulations}
\title{Run Simulations for the Settings in Carlotti and Parast (2026)}
\usage{
Carlotti_and_Parast_2026_simulations(seed, n_simulations, parallel = TRUE)
}
\arguments{
\item{seed}{Numeric. A random seed for reproducibility of the simulations.}

\item{n_simulations}{Numeric. The number of simulation runs to execute for the setting.}

\item{parallel}{Logical. Whether to run the simulations in parallel or sequentially (default is TRUE).}
}
\value{
A data frame containing:
\itemize{
  \item \code{setting}: The index of the simulation setting (1 in this case).
  \item \code{simulation}: The individual simulation run ID.
  \item \code{Bayesian_epsilon}, \code{frequentist_epsilon}: Numeric. Surrogate validation thresholds for the Bayesian and frequentist tests, respectively.
  \item \code{Bayesian_coverage}, \code{frequentist_coverage}: Boolean. Indicates whether the true surrogate effect is within the credible or confidence interval for the Bayesian and frequentist tests, respectively.
  \item \code{Bayesian_power}, \code{frequentist_power}: Boolean. Indicates whether the upper bound of the credible or confidence interval is below the threshold for the Bayesian and frequentist tests, respectively.
  \item \code{n_sample}, \code{n_simulations}, \code{n_chains}, \code{n_iterations}: Numeric. Metadata about the simulation parameters.
  \item \code{timestamp}: Character. Date and time when the simulation was completed.
  \item \code{seed}: Numeric. The random seed used for reproducibility.
}
}
\description{
This function executes the simulation study described in Carlotti and Parast
(2026) for the covariate-augmented setting with a binary covariate \eqn{X}. It
performs the frequentist and Bayesian surrogate evaluation tests from Parast
et al. (2024) and Carlotti and Parast (2026), respectively, and returns a
structured data frame of the results. The execution is parallelized using the
\code{future} framework.
}
\details{
The function runs a total of 500 simulations for this setting, generating
data using \code{DGP_X} and processes the results using \code{BSET_X}. The
simulation utilizes MCMC sampling via \code{rstan} for the Bayesian
estimation components. Note that it relies on an external object 
\code{estimands_Carlotti_and_Parast_2026} that contains the true values of
the computed using the function
\code{compute_estimands_Carlotti_and_Parast_2026}.
}
