% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_estimands_Carlotti_and_Parast_2026.R
\name{compute_estimands_Carlotti_and_Parast_2026}
\alias{compute_estimands_Carlotti_and_Parast_2026}
\title{Monte Carlo Computation of the Estimands for the Simulation Study in Carlotti and Parast (2026)}
\usage{
compute_estimands_Carlotti_and_Parast_2026(MC_samples)
}
\arguments{
\item{MC_samples}{Integer. The number of Monte Carlo samples to generate per setting.}
}
\value{
A data frame containing the Monte Carlo estimates for each setting:
\itemize{
  \item \code{setting}: The index of the simulation setting.
  \item \code{U_Y_MC}, \code{U_S_MC}, \code{delta_MC}: Parameters of interest from Parast et al. (2024).
  \item \code{V_Y_MC}, \code{V_S_MC}, \code{theta_MC}: Parameters of interest from Carlotti and Parast (2026).
}
}
\description{
This function iterates through the simulation settings defined in Carlotti
and Parast (2026) and estimates the true values of \eqn{U_Y}, \eqn{U_S},
\eqn{\delta}, \eqn{V_Y}, \eqn{V_S}, and \eqn{\theta} using a Monte Carlo
dataset generated according to the specified data-generating processes. The
settings are defined as follows:
\itemize{
  \item Setting 1: \strong{X binary}
  \itemize{
    \item If \eqn{X = 1}:
    \deqn{Y_1 \sim \mathcal{N}(5, 1), \quad Y_0 \sim \mathcal{N}(0, 1)}
    \deqn{S_1 = Y_1 + \mathcal{N}(0, 1), \quad S_0 = Y_0 + \mathcal{N}(0, 1)}
    \item If \eqn{X = 0}:
    \deqn{Y_1 \sim \mathcal{N}(5, 1), \quad Y_0 \sim \mathcal{N}(0, 1)}
    \deqn{S_1 = Y_1 + \mathcal{N}(-10, 1), \quad S_0 = Y_0 + \mathcal{N}(-10, 1)}
  }
  \item Setting 2: \strong{X Gaussian}
  \deqn{()}
}
}
