% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estimands_Parast_et_al_2024}
\alias{estimands_Parast_et_al_2024}
\title{True estimands for Parast et al. (2024)}
\format{
A data frame with 4 rows and 7 columns:
\describe{
  \item{setting}{The index of the simulation setting.}
  \item{U_Y_MC}{Numeric vector of Monte Carlo estimates.}
  \item{U_S_MC}{Numeric vector of Monte Carlo estimates.}
  \item{delta_MC}{Numeric vector of Monte Carlo estimates.}
  \item{V_Y_MC}{Numeric vector of Monte Carlo estimates.}
  \item{V_S_MC}{Numeric vector of Monte Carlo estimates.}
  \item{theta_MC}{Numeric vector of Monte Carlo estimates.}
}
}
\source{
Computed using the function \code{compute_estimands_Parast_et_al_2024(1000000)}.
}
\usage{
estimands_Parast_et_al_2024
}
\description{
A dataset containing the Monte Carlo estimates of the parameters of interest
for the simulation settings considered in Parast et al. (2024).
}
\keyword{datasets}
