% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Parast_et_al_2024_simulations_grid}
\alias{Parast_et_al_2024_simulations_grid}
\title{Simulation grid for Parast et al. (2024)}
\format{
A data frame with multiple columns:
\describe{
  \item{setting}{The index of the simulation setting.}
  \item{simulation}{The index of the simulation run.}
  \item{n_sample}{The sample size used.}
  \item{n_chains}{The number of MCMC chains used.}
  \item{n_iterations}{The number of MCMC iterations.}
  \item{n_simulations}{The total number of simulations.}
  \item{timestamp}{The timestamp of execution.}
  \item{seed}{The random seed used.}
  \item{BF_alternative}{The alternative hypothesis used for the Bayes factor.}
  \item{Bayesian_epsilon}{The threshold for the Bayesian test.}
  \item{frequentist_epsilon}{The threshold for the frequentist test.}
  \item{Bayesian_CI_upper}{The upper bound of the Bayesian credible interval.}
  \item{frequentist_CI_upper}{The upper bound of the frequentist confidence interval.}
  \item{Bayesian_coverage}{Logical. Indicates if the true value is in the credible interval.}
  \item{frequentist_coverage}{Logical. Indicates if the true value is in the confidence interval.}
  \item{Bayesian_power}{Logical. Indicates if the Bayesian test rejected the null.}
  \item{frequentist_power}{Logical. Indicates if the frequentist test rejected the null.}
}
}
\source{
Generated using the code in the \code{simulations} folder of the
  BSET GitHub repository (\url{https://github.com/PietroCarlotti/BSET}).
}
\usage{
Parast_et_al_2024_simulations_grid
}
\description{
A dataset containing the grid of simulation results for Parast et al. (2024).
}
\keyword{datasets}
