% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP_no_X.R
\name{DGP_no_X}
\alias{DGP_no_X}
\title{Data Generating Process without Baseline Covariates}
\usage{
DGP_no_X(
  n,
  p,
  mu_star = NULL,
  Sigma_star = NULL,
  model = c("Gaussian", "misspecified")
)
}
\arguments{
\item{n}{Integer. Total sample size.}

\item{p}{Numeric. Probability of being assigned to the treatment group (Z=1).}

\item{mu_star}{Numeric vector. The mean vector for \eqn{P}. Required if \code{model = "Gaussian"}.}

\item{Sigma_star}{Matrix. The covariance matrix for \eqn{P}. Required if \code{model = "Gaussian"}.}

\item{model}{Character. The type of data generation: \code{"Gaussian"} 
or \code{"misspecified"}.}
}
\value{
A list containing:
\itemize{
  \item \code{Z}: Treatment assignment vector.
  \item \code{n1}: Number of treated units.
  \item \code{n0}: Number of control units.
  \item \code{P}: Full matrix of potential outcomes.
  \item \code{P_observed}: Observed outcomes \eqn{(Y, S)} corresponding to the assigned treatment \eqn{Z}.
  \item \code{P_unobserved}: Counterfactual outcomes under the opposite treatment.
}
This function is not a primary user-facing function of the package and
does not include examples.
}
\description{
This function generates potential outcomes from the simulation settings
described in Parast et al. (2024). It creates a dataset of potential outcomes
\deqn{P = (Y_1, S_1, Y_0, S_0)} and observed outcomes 
\deqn{P_{observed} = (Y, S)} based on a random treatment assignment \eqn{Z}.
}
\details{
The function supports two types of data-generating processes:
\itemize{
  \item \strong{Gaussian model:} Potential outcomes are drawn from a 
  multivariate normal distribution:
  \deqn{P \sim \mathcal{N}_{4}(\mu^{*}, \Sigma^{*}).}
  \item \strong{Non-linear model:} Potential outcomes for the surrogate are generated from a non-Gaussian
  distribution, and the potential outcomes for the primary outcome are generated from a non-linear function of the surrogate plus non-Gaussian noise.
}
}
