% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP_X.R
\name{DGP_X_Gaussian}
\alias{DGP_X_Gaussian}
\title{Data Generating Process with a Continuous Covariate}
\usage{
DGP_X_Gaussian(n, p, beta, Sigma, m, s)
}
\arguments{
\item{n}{Integer. Total sample size.}

\item{p}{Numeric. Probability of being assigned to the treatment group \eqn{(Z=1)}.}

\item{beta}{Numeric vector. Coefficients for the linear function of the mean vector.}

\item{Sigma}{Matrix. Covariance matrix for the potential outcomes.}

\item{m}{Numeric. Mean of the continuous covariate X.}

\item{s}{Numeric. Standard deviation of the continuous covariate X.}
}
\value{
A list containing:
\itemize{
 \item \code{X}: The continuous covariate vector.
 \item \code{Z}: Treatment assignment vector.
 \item \code{n1}: Number of treated units.
 \item \code{n0}: Number of control units.
 \item \code{P}: Full matrix of potential outcomes.
 \item \code{P_observed}: Observed outcomes \eqn{(Y, S)} corresponding to the assigned treatment \eqn{Z}.
 \item \code{P_unobserved}: Counterfactual outcomes under the opposite treatment.
}
}
\description{
This function generates potential outcomes from a data generating process
similar to the one described in Parast et al. (2024), but with the addition
of a continuous covariate X. It creates a dataset of potential outcomes
\deqn{P = (Y_1, S_1, Y_0, S_0)} and observed outcomes
\deqn{P_{observed} = (Y, S)} based on a random treatment assignment \eqn{Z}.
}
\details{
The potential outcomes are generated from multivariate normal
distributions with mean vector and covariance matrix that depend on the
value of \eqn{X}. Specifically, the mean vector is a linear function of \eqn{X}:
\deqn{\mu(X) = x \cdot (\beta_{Y1}, \beta_{S1}, \beta_{Y0}, \beta_{S0})^T,}
and the covariance matrix is constant across values of \eqn{X}:
\deqn{\Sigma(X) = \Sigma.}
}
