% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estimands_Carlotti_and_Parast_2026}
\alias{estimands_Carlotti_and_Parast_2026}
\title{True estimands for Carlotti and Parast (2026)}
\format{
A data frame with 1 row and 6 columns:
\describe{
  \item{U_Y_MC}{Numeric vector of Monte Carlo estimates.}
  \item{U_S_MC}{Numeric vector of Monte Carlo estimates.}
  \item{delta_MC}{Numeric vector of Monte Carlo estimates.}
  \item{V_Y_MC}{Numeric vector of Monte Carlo estimates.}
  \item{V_S_MC}{Numeric vector of Monte Carlo estimates.}
  \item{theta_MC}{Numeric vector of Monte Carlo estimates.}
}
}
\source{
Computed using the function \code{compute_estimands_Carlotti_and_Parast_2026(1000000)}.
}
\usage{
estimands_Carlotti_and_Parast_2026
}
\description{
A dataset containing the Monte Carlo estimates of the parameters of interest
for the simulation settings considered in Carlotti and Parast (2026).
}
\keyword{datasets}
